1 - Segmentation/Run-UNET/requirements.txt
numpy
pandas
seaborn
matplotlib
torch
torchvision
tqdm
opencv-python
scikit-learn
pyyaml
